# Speed up mdtests