@MichaReiser 2024-10-20
The positive. This shows that Red Knot's incremental computation works :)