@MichaReiser 2024-10-20
Yes, isolation is guaranteed for as long as the test framework correctly resets the DB state after each run and salsa has no caching bugs.