@MichaReiser 2024-10-20
The most ideal solution would be to prewarm a static db and then deep clone it for every test but salsa doesn't allow deep cloning the storage. Maybe that's something we can do once we have persistent caching