@MichaReiser 2024-10-21
### Settings support

I don't see this as an issue. Settings will be salsa inputs and changing them invalidates the relevant queries. We already support some settings today and there are updates to update them: `Program::update_search_paths` and `Program::set_target_version`. It may require exposing some additional APIs but most should be sharable with the code used by the CLI watch mode and the LSP. 

Having said that, I'm not sure if we should allow changing all settings between tests in the same test suite. For example, we could  consider restricting the target version to apply to the entire test suite because changing the version for each test leads to large cache invalidation and slower wall time. But that also depends on how such a restriction limits the test framework ergonocmis.