@carljm 2024-10-21
> Settings will be salsa inputs and changing them invalidates the relevant queries.

That sounds great! My memory wasn't clear on how we ended up handling all the settings :)