@carljm 2024-10-21
> I'm not sure if we should allow changing all settings between tests in the same test suite... But that also depends on how such a restriction limits the test framework ergonomics.

Yeah, I agree we'll need to see how this works out. I'm open to having some such restrictions to help us keep the test suite faster, but I think there likely may be cases where it makes the tests a lot more readable to have exactly-parallel tests for behavior in different Python versions live next to each other in the same file, rather than in separate files.