@MichaReiser 2024-10-21
Shoot. I made a title and then forgot to cover your second question. 

My preference is to add a environment variable that allows filtering by test name. This way we get two functionalities with one env variable:

* You can set it to debug a specific test 
* You can set it see if the test passes in isolation.

We probably also want to setup tracing by calling into `ruff_db::testing::setup_logging`. I create two new issues